third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-in-worker-expected.txt

CONSOLE ERROR: line 3: Unhandled promise rejection
CONSOLE ERROR: line 21: Unhandled promise rejection
Tests that uncaught promise rejections happenned in workers are logged into console.

worker-with-unhandled-promises.js:3 Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: err1}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: err1
worker-with-unhandled-promises.js:21 Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: err2}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: err2
